refactor(dh): move files around
