--- libs/dh/reports/data-access-graphql/.eslintrc.json
@@ -1,4 +1,0 @@
-{
-  "extends": ["../../../../.eslintrc.json"],
-  "ignorePatterns": ["!**/*"]
-}
