--- libs/dh/reports/data-access-graphql/project.json
@@ -1,14 +1,0 @@
-{
-  "name": "dh-reports-data-access-graphql",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "library",
-  "sourceRoot": "libs/dh/reports/data-access-graphql",
-  "prefix": "dh",
-  "tags": ["product:dh", "domain:reports", "type:data-access"],
-  "implicitDependencies": ["api-dh"],
-  "targets": {
-    "lint": {
-      "executor": "@nx/eslint:lint"
-    }
-  }
-}
