+++ libs/dh/reports/feature-settlement-reports/src/components/report-details/duration.component.ts
