+++ libs/dh/reports/feature-settlement-reports/src/components/util/dh-is-period-one-full-month.ts
